new-arduino: improve architecture
This is a broad-sweeping commit for the new-arduino codebase.
In the client, result types have been unified,
there are the beginnings a getopt() -based CLI,
and the listener thread was relieved of its decoding and output duties,
which have been relegated to a new 'monitor' thread.
Also added are macros for panicking.

Unfortunately, the interaction between the threads is seriously f#$%ed up
because I haven't really done much multithreaded programming before,
much less with POSIX Threads.
I think what needs to happen is,
the listener and monitor threads should not be managed directly in main()
but abstracted behind a adptc_listen_system structure or such.
There will be an adptc_listen_start() function
that initializes this structure
and recieves a pointer to the monitor thread function
and an arbitrary pointer to be passed to the monitor thread
and creates the listen and monitor threads in the correct order.
There will also be an adptc_listen_stop() function
that exits the listen and monitor threads in the correct order
and cleans up all managed resources.